Separate `Extractor`s from framework responsibilities (captures)
Extractors now return an `Option` (a variant of `Optional`
accepting nulls) instead of the previously returned `Match`.

Match received a `Captures` field and reduced the number
of its factory methods to 2 (as one would expect from a
variant of `Optional`).

`Captures` is an immutable head-tail list based typesafe
heterogenic map, with `O(N)` `get` and `addAll` operations.
This should be enough given the usual size of Matchers and should
even save some allocations when compared to an implementation
 based on `unmodifiableList(ArrayList)` or similar with Maps.